chore: address PR comments about potential circular dependencies
