+++ src/Signin.tsx
@@ -17,7 +17,7 @@ import {
   setToLocalStorage,
 } from 'src/localStorage'
 import {isFlagEnabled} from 'src/shared/utils/featureFlag'
-import {getPublicFlags} from 'src/shared/actions/flags'
+import {getPublicFlags} from 'src/shared/thunks/flags'
 
 // Actions
 import {notify as notifyAction} from 'src/shared/actions/notifications'
